refactor: modularize checkout helpers
